add image command
