+++ core/core/config_scan.go
