+++ core/meta/datastructures/v1/image_scan.go
@@ -1,0 +1,7 @@
+package v1
+
+type ImageScanInfo struct {
+	Username string
+	Password string
+	Image    string
+}
