+++ core/meta/ksinterface.go
@@ -3,6 +3,7 @@ package meta
 import (
 	"context"
 
+	"github.com/anchore/grype/grype/presenter/models"
 	"github.com/kubescape/kubescape/v2/core/cautils"
 	metav1 "github.com/kubescape/kubescape/v2/core/meta/datastructures/v1"
 	"github.com/kubescape/kubescape/v2/core/pkg/resultshandling"
@@ -25,4 +26,7 @@ type IKubescape interface {
 
 	// patch
 	Patch(ctx context.Context, patchInfo *metav1.PatchInfo) error
+
+	// scan image
+	ScanImage(ctx context.Context, imgScanInfo *metav1.ImageScanInfo, scanInfo *cautils.ScanInfo) (*models.PresenterConfig, error)
 }
